pyproject.toml
[project]
name = "epm-note-engine"
version = "0.1.0"
description = "EPM Note Engine - AI-powered article generation system for Note.com"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "EPM Team"}
]

dependencies = [
    # Web Framework
    "streamlit>=1.30.0",

    # AI/LLM
    "langchain>=0.1.0",
    "langgraph>=0.2.0,<0.3.0",
    "anthropic>=0.18.0",
    "openai>=1.10.0",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",

    # Vector Store
    "chromadb>=0.4.22",

    # Search API
    "tavily-python>=0.3.0",

    # Browser Automation
    "playwright>=1.41.0",

    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
